requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.0
requests==2.31.0
openai==1.8.0
gunicorn==20.1.0